Feat: create useLoginStatus custom hook and refactor App
